mulecube-controlpanel-user/hw-monitor/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
smbus2==0.4.3
pydantic==2.5.3
python-dotenv==1.0.0
